CN_solutions/im2recipe/requirements.txt
word2vec
numpy
scipy==1.1.0
pillow
lmdb
scikit-learn
tqdm
nltk
simplejson
torch
torchfile
torchvision
torchwordemb
pickle
pymysql
pymilvus==1.0.1
fastapi
uvicorn
python-multipart
